#35 [FEAT] 프로필 수정 기능 구현

+++ saju-backend/src/main/java/com/saju/sajubackend/api/filter/dto/UpdateProfileRequest.java
@@ -12,5 +12,5 @@ public class UpdateProfileRequest {
     private String smoking;
     private String drinking;
     private Integer height;
-    private Integer cityCode;
+    private Long cityCode;
 }
